pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synthetic-data-seeder"
version = "0.1.0"
description = "AI-Assisted Universal Synthetic Data & Seeder Generator"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "agno>=2.0",
    "openai>=1.0",
    "sqlparse>=0.4",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "pypdf>=4.0",
]

[project.scripts]
synthetic-seeder = "synthetic_seeder.cli:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff"]
web = ["fastapi>=0.109", "uvicorn[standard]>=0.27"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
